Lecture 25 down
Hooked up user login.  Modified the way they implemented the
passport.auth method.  The course calls it inside of the route.post
function, but a little research showed me that I could just put it as
middleware for the router and achieve the same effect without having to
manually catch and pass all the errors to next.  Also, the user name
prints to the main screen now as proof that login was a success.